requirements.txt
Django==4.0.1
djangorestframework==3.11.0
psycopg2-binary==2.8.6
requests==2.28.0
black==22.3.0
isort==5.10.1
django-filter==2.3